+++ NEWS.md
@@ -1,6 +1,7 @@
 # crew 0.8.0.9001 (development)
 
 * Require `nanonext` >= 0.12.0 and `mirai` >= 0.12.0.
+* Return to always re-launching backlogged inactive workers (#79, https://github.com/shikokuchuo/mirai/discussions/95).
 
 # crew 0.8.0